Extract mutaions form store file
